% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{friends_lines}
\alias{friends_lines}
\title{Dataset for TV series friends}
\format{A data frame with 64280 rows and 6 variables:
\describe{
  \item{episodes_id}{the id of episodes, integer with 4 digits, first two are for season, latter two are for episode number}
  \item{person}{the name of the speaker of the line, character}
  \item{line}{a single line of the speaker, character}
  \item{type}{type of the line, 'person' or 'scene'}
  \item{season}{the season that the line appears}
  \item{scene2}{the scene the line appears, integer}
}}
\source{
\url{https://fangj.github.io/friends/}
}
\usage{
friends_lines
}
\description{
This dataset is built from TV series transcripts friends. It contains the lines each character speak, name of the speakers,
scene number when the dialogue take places.
}
\keyword{datasets}
