% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{US_price_month}
\alias{US_price_month}
\title{Monthly Official Price index and online price index in US from 2008.07-2015.08}
\format{An object of class \code{data.frame} with 86 rows and 8 columns.}
\source{
\url{https://dataverse.harvard.edu/file.xhtml?persistentId=doi:10.7910/DVN/6RQCRS/KGU8B0&version=2.0}
        \url{https://www.bls.gov/developers/api_signature_v2.htm}
}
\usage{
US_price_month
}
\description{
A dataset containing monthly the Official price index leased by Bureau of labor statistics
along with the price index built in the paper ‘THE BILLION PRICES PROJECT: USING
ONLINE PRICES FOR MEASUREMENT AND RESEARCH’
}
\keyword{datasets}
